+++ financius/src/main/java/com/code44/finance/modules/ViewModule.java
@@ -1,8 +1,8 @@
 package com.code44.finance.modules;
 
+import com.code44.finance.ui.common.ActiveIntervalView;
+import com.code44.finance.ui.reports.categories.CategoriesReportView;
 import com.code44.finance.views.AccountsView;
-import com.code44.finance.views.ActiveIntervalView;
-import com.code44.finance.views.CategoriesReportView;
 import com.code44.finance.views.OverviewGraphView;
 
 import dagger.Module;